stdb: new (untested!) "soft-sync" tdb wrapper.
This provides an alternative to fsync() for tdb, by weakening
the "durability" part of ACID.  It rarely does an fsync, instead
it keeps a replay log with checksums.  This means that in a crash,
older transactions may be forgotten, but the database will be
consistent.

This may be an acceptable tradeoff for ctdb, where an entire cluster
crash is unlikely.  It remains to be seen if the overhead of writev()
outweighs the overhead of the three fsync().

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>